frontend client place order

+++ src/main/java/com/g4t2project/g4t2project/DTO/cleaningTaskDTO.java
@@ -1,23 +1,23 @@
 package com.g4t2project.g4t2project.DTO;
 
 public class cleaningTaskDTO {
-    private int propertyId;
+    private Long propertyId;
     private String shift;
     private String date;
     private boolean acknowledged;
     
-    public cleaningTaskDTO(int propertyId, String shift, String date, boolean acknowledged) {
+    public cleaningTaskDTO(Long propertyId, String shift, String date, boolean acknowledged) {
         this.propertyId = propertyId;
         this.shift = shift;
         this.date = date;
         this.acknowledged = acknowledged;
     }   
 
-    public int getPropertyId() {    
+    public Long getPropertyId() {    
         return propertyId;
     }
 
-    public void setPropertyId(int propertyId) {
+    public void setPropertyId(Long propertyId) {
         this.propertyId = propertyId;
     }
